Add self properties to chains
This is to allow ontologists to change the 'root' used for nominal
expressions. If 'it' is added, such as in 'the balance of it', then the
nominal variable is added relative to the current node of the restriction
tree. Otherwise, it is added from the root of the class expression.

Fixed: a bug with overeager matches in duration literals